+++ packages/cojson/src/crypto/PureJSCrypto.ts
@@ -193,7 +193,9 @@ export class PureJSCrypto extends CryptoProvider<Blake3State> {
     try {
       return JSON.parse(textDecoder.decode(plaintext));
     } catch (e) {
-      logger.error("Failed to decrypt/parse sealed message", e);
+      logger.error(
+        "Failed to decrypt/parse sealed message: " + (e as Error)?.message,
+      );
       return undefined;
     }
   }