+++ packages/cojson/src/crypto/crypto.ts
@@ -160,7 +160,7 @@ export abstract class CryptoProvider<Blake3State = any> {
     try {
       return parseJSON(this.decryptRaw(encrypted, keySecret, nOnceMaterial));
     } catch (e) {
-      logger.error("Decryption error", e);
+      logger.error("Decryption error: " + (e as Error)?.message);
       return undefined;
     }
   }