+++ examples/chat-rn/CHANGELOG.md
@@ -1,5 +1,35 @@
 # chat-rn
 
+## 1.0.53
+
+### Patch Changes
+
+- Updated dependencies [7491711]
+  - jazz-tools@0.9.15
+  - jazz-react-native@0.9.15
+
+## 1.0.52
+
+### Patch Changes
+
+- Updated dependencies [3df93cc]
+  - jazz-tools@0.9.14
+  - jazz-react-native@0.9.14
+
+## 1.0.51
+
+### Patch Changes
+
+- jazz-react-native@0.9.13
+- jazz-tools@0.9.13
+
+## 1.0.50
+
+### Patch Changes
+
+- jazz-react-native@0.9.12
+- jazz-tools@0.9.12
+
 ## 1.0.49
 
 ### Patch Changes