+++ examples/clerk/CHANGELOG.md
@@ -1,5 +1,39 @@
 # minimal-auth-clerk
 
+## 0.0.35
+
+### Patch Changes
+
+- Updated dependencies [7491711]
+  - jazz-tools@0.9.15
+  - jazz-react@0.9.15
+  - jazz-react-auth-clerk@0.9.15
+
+## 0.0.34
+
+### Patch Changes
+
+- Updated dependencies [3df93cc]
+  - jazz-tools@0.9.14
+  - jazz-react@0.9.14
+  - jazz-react-auth-clerk@0.9.14
+
+## 0.0.33
+
+### Patch Changes
+
+- jazz-react@0.9.13
+- jazz-react-auth-clerk@0.9.13
+- jazz-tools@0.9.13
+
+## 0.0.32
+
+### Patch Changes
+
+- jazz-react@0.9.12
+- jazz-react-auth-clerk@0.9.12
+- jazz-tools@0.9.12
+
 ## 0.0.31
 
 ### Patch Changes