+++ examples/inspector/CHANGELOG.md
@@ -1,5 +1,21 @@
 # jazz-example-inspector
 
+## 0.0.98
+
+### Patch Changes
+
+- Updated dependencies [8d29e50]
+  - cojson-transport-ws@0.9.13
+  - cojson@0.9.13
+
+## 0.0.97
+
+### Patch Changes
+
+- Updated dependencies [15d4b2a]
+  - cojson-transport-ws@0.9.12
+  - cojson@0.9.12
+
 ## 0.0.96
 
 ### Patch Changes