+++ examples/music-player/CHANGELOG.md
@@ -1,5 +1,39 @@
 # jazz-example-musicplayer
 
+## 0.0.56
+
+### Patch Changes
+
+- Updated dependencies [7491711]
+  - jazz-tools@0.9.15
+  - jazz-inspector@0.9.15
+  - jazz-react@0.9.15
+
+## 0.0.55
+
+### Patch Changes
+
+- Updated dependencies [3df93cc]
+  - jazz-tools@0.9.14
+  - jazz-inspector@0.9.14
+  - jazz-react@0.9.14
+
+## 0.0.54
+
+### Patch Changes
+
+- jazz-inspector@0.9.13
+- jazz-react@0.9.13
+- jazz-tools@0.9.13
+
+## 0.0.53
+
+### Patch Changes
+
+- jazz-inspector@0.9.12
+- jazz-react@0.9.12
+- jazz-tools@0.9.12
+
 ## 0.0.52
 
 ### Patch Changes