+++ examples/todo/CHANGELOG.md
@@ -1,5 +1,35 @@
 # jazz-example-todo
 
+## 0.0.152
+
+### Patch Changes
+
+- Updated dependencies [7491711]
+  - jazz-tools@0.9.15
+  - jazz-react@0.9.15
+
+## 0.0.151
+
+### Patch Changes
+
+- Updated dependencies [3df93cc]
+  - jazz-tools@0.9.14
+  - jazz-react@0.9.14
+
+## 0.0.150
+
+### Patch Changes
+
+- jazz-react@0.9.13
+- jazz-tools@0.9.13
+
+## 0.0.149
+
+### Patch Changes
+
+- jazz-react@0.9.12
+- jazz-tools@0.9.12
+
 ## 0.0.148
 
 ### Patch Changes