+++ packages/cojson-storage-indexeddb/CHANGELOG.md
@@ -1,5 +1,21 @@
 # cojson-storage-indexeddb
 
+## 0.9.13
+
+### Patch Changes
+
+- Updated dependencies [8d29e50]
+  - cojson-storage@0.9.13
+  - cojson@0.9.13
+
+## 0.9.12
+
+### Patch Changes
+
+- Updated dependencies [15d4b2a]
+  - cojson-storage@0.9.12
+  - cojson@0.9.12
+
 ## 0.9.11
 
 ### Patch Changes