+++ packages/cojson-storage-sqlite/CHANGELOG.md
@@ -1,5 +1,23 @@
 # cojson-storage-sqlite
 
+## 0.9.13
+
+### Patch Changes
+
+- 8d29e50: Restore the logger wrapper and adapt the API to pino
+- Updated dependencies [8d29e50]
+  - cojson-storage@0.9.13
+  - cojson@0.9.13
+
+## 0.9.12
+
+### Patch Changes
+
+- 15d4b2a: Revert the custom logger
+- Updated dependencies [15d4b2a]
+  - cojson-storage@0.9.12
+  - cojson@0.9.12
+
 ## 0.9.11
 
 ### Patch Changes