+++ packages/cojson/CHANGELOG.md
@@ -1,5 +1,17 @@
 # cojson
 
+## 0.9.13
+
+### Patch Changes
+
+- 8d29e50: Restore the logger wrapper and adapt the API to pino
+
+## 0.9.12
+
+### Patch Changes
+
+- 15d4b2a: Revert the custom logger
+
 ## 0.9.11
 
 ### Patch Changes