+++ packages/jazz-react/CHANGELOG.md
@@ -1,5 +1,43 @@
 # jazz-react
 
+## 0.9.15
+
+### Patch Changes
+
+- Updated dependencies [7491711]
+  - jazz-tools@0.9.15
+  - jazz-browser@0.9.15
+  - jazz-react-core@0.8.56
+
+## 0.9.14
+
+### Patch Changes
+
+- Updated dependencies [3df93cc]
+  - jazz-tools@0.9.14
+  - jazz-browser@0.9.14
+  - jazz-react-core@0.8.55
+
+## 0.9.13
+
+### Patch Changes
+
+- Updated dependencies [8d29e50]
+  - cojson@0.9.13
+  - jazz-browser@0.9.13
+  - jazz-react-core@0.8.54
+  - jazz-tools@0.9.13
+
+## 0.9.12
+
+### Patch Changes
+
+- Updated dependencies [15d4b2a]
+  - cojson@0.9.12
+  - jazz-browser@0.9.12
+  - jazz-react-core@0.8.53
+  - jazz-tools@0.9.12
+
 ## 0.9.11
 
 ### Patch Changes